+++ CHANGES.rst
@@ -1,5 +1,15 @@
 .. currentmodule:: dirty_cat
 
+Release 0.5.0
+==============
+
+Major changes
+-------------
+
+* New experimental feature: :class:`FeatureAugmenter`, a transformer
+  that augments with :func:`fuzzy_join` the number of features in a main table
+  by using information from auxilliary tables.
+
 Release 0.4.0
 ==============
 
@@ -8,9 +18,11 @@ Major changes
 
 * New experimental feature: joining tables using :func:`fuzzy_join` by approximate key matching. Matches are based
   on string similarities and the nearest neighbors matches are found for each category.
+
 * **datasets.fetching**: contains a new function :func:`fetch_world_bank_indicator` that can be
-    used to download any indicator from the World Bank Open Data platform. It only needs the
-    indicator ID that can be found on the website.
+  used to download any indicator from the World Bank Open Data platform. It only needs the
+  indicator ID that can be found on the website.
+
 * Unnecessary API has been made private: everything (files, functions, classes)
   starting with an underscore shouldn't be imported in your code.